+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Table: `wideFirstColumn` is now `false` by default
 - Tag: `angled` prop is removed
 - propTypes are removed in favor of TypeScript
+- Button `text` prop now behaves in the same way as `inline`, it removes paddings and borders and inherits the font size and line height from parent. The `text` prop is deprecated in favor of `inline`. `inline` prop is false by default for buttons without icons and true for buttons with icons.
 
 ## [6.0.0](https://youtrack.jetbrains.com/issue/RG-2303/Release-Ring-UI-6.0)